Update and clean up based on @kbenne comments
