+++ tst/EnergyPlus/unit/Fixtures/HVACFixture.hh
@@ -49,8 +49,6 @@ namespace EnergyPlus {
 		}
 	};
 
-	typedef HVACFixture HVACDeathTestFixture;
-
 }
 
 #endif
