+++ tst/EnergyPlus/unit/Fixtures/IdfParserFixture.hh
@@ -34,50 +34,41 @@ namespace EnergyPlus {
 		}
 
 		void eat_whitespace( std::string const & idf, size_t & index ) {
-			IdfParser parser;
-			parser.eat_whitespace( idf, index );
+			IdfParser::eat_whitespace( idf, index );
 		}
 
 		void eat_comment( std::string const & idf, size_t & index ) {
-			IdfParser parser;
-			parser.eat_comment( idf, index );
+			IdfParser::eat_comment( idf, index );
 		}
 
 		std::string parse_string( std::string const & idf, size_t & index, bool & success) {
-			IdfParser parser;
-			return parser.parse_string( idf, index, success );
+			return IdfParser::parse_string( idf, index, success );
 		}
 
 		std::string parse_value( std::string const & idf, size_t & index, bool & success) {
-			IdfParser parser;
-			return parser.parse_value( idf, index, success );
+			return IdfParser::parse_value( idf, index, success );
 		}
 
 		IdfParser::Token look_ahead( std::string const & idf, size_t index)
 		{
-			IdfParser parser;
-			return parser.look_ahead( idf, index );
+			return IdfParser::look_ahead( idf, index );
 		}
 
 		IdfParser::Token next_token( std::string const & idf, size_t & index)
 		{
-			IdfParser parser;
-			return parser.next_token( idf, index );
+			return IdfParser::next_token( idf, index );
 		}
 
 		std::vector< std::vector< std::string > > parse_idf( std::string const & idf, size_t & index, bool & success ) {
-			IdfParser parser;
-			return parser.parse_idf( idf, index, success );
+			return IdfParser::parse_idf( idf, index, success );
 		}
 
 		std::vector< std::string > parse_object( std::string const & idf, size_t & index, bool & success ) {
-			IdfParser parser;
-			return parser.parse_object( idf, index, success );
+			return IdfParser::parse_object( idf, index, success );
 		}
 		
 	};
 
-	typedef IdfParserFixture IdfParserDeathTestFixture;
 }
 
 #endif
