+++ tst/EnergyPlus/unit/Fixtures/SQLiteFixture.hh
@@ -155,8 +155,6 @@ namespace EnergyPlus {
 		}
 	};
 
-	typedef SQLiteFixture SQLiteDeathTestFixture;
-
 }
 
 #endif
