+++ tst/EnergyPlus/unit/Humidifiers.unit.cc
@@ -6,6 +6,9 @@
 // ObjexxFCL Headers
 
 // EnergyPlus Headers
+#include <EnergyPlus/CurveManager.hh>
+#include <EnergyPlus/DataEnvironment.hh>
+#include <EnergyPlus/DataGlobals.hh>
 #include <EnergyPlus/Humidifiers.hh>
 
 #include "Fixtures/HVACFixture.hh"
@@ -20,10 +23,7 @@ using namespace EnergyPlus::CurveManager;
 
 namespace EnergyPlus {
 
-	typedef HVACFixture HumidifiersFixture;
-	typedef HVACFixture HumidifiersDeathTestFixture;
-
-	TEST_F( HumidifiersFixture, Sizing ) {
+	TEST_F( HVACFixture, Humidifiers_Sizing ) {
 		SysSizingRunDone = true;
 		CurSysNum = 1;
 		NumElecSteamHums = 0;
@@ -58,7 +58,7 @@ namespace EnergyPlus {
 
 	}
 
-	TEST_F( HumidifiersFixture, AutoSizing ) {
+	TEST_F( HVACFixture, Humidifiers_AutoSizing ) {
 		SysSizingRunDone = true;
 		CurSysNum = 1;
 		NumElecSteamHums = 0;
@@ -97,7 +97,7 @@ namespace EnergyPlus {
 		EXPECT_NEAR( 265257.67, thisHum.NomPower, 1.0E-02 ); // Watts
 	}
 
-	TEST_F( HumidifiersFixture, EnergyUse ) {
+	TEST_F( HVACFixture, Humidifiers_EnergyUse ) {
 		HumidifierData thisHum;
 
 		TimeStepSys = 0.25;
@@ -148,7 +148,7 @@ namespace EnergyPlus {
 		EXPECT_DOUBLE_EQ( 93339384.987223208, thisHum.GasUseEnergy );
 	}
 
-	TEST_F( HumidifiersFixture, GetHumidifierInput ) {
+	TEST_F( HVACFixture, Humidifiers_GetHumidifierInput ) {
 		std::string const idf_objects = delimited_string({
 			"Version,8.3;",
 			"Humidifier:Steam:Gas,",
@@ -185,7 +185,7 @@ namespace EnergyPlus {
 
 	}
 
-	TEST_F( HumidifiersFixture, ThermalEfficiency ) {
+	TEST_F( HVACFixture, Humidifiers_ThermalEfficiency ) {
 		// tests thermal efficiency modifier curve use
 
 		HumidifierData thisHum;
